[Change] separate admin routes from index.js
